+++ src/extensions/mvs/helpers/animation.ts
@@ -403,8 +403,7 @@ function makePaletteFunctionContinuous(props: MVSContinuousPaletteProps): (value
             return overflowColor;
         }
         const q = (x - checkpoints[gteIdx - 1]) / (checkpoints[gteIdx] - checkpoints[gteIdx - 1]);
-        // TODO: consider HSL interpolation
-        return Color.interpolate(colors[gteIdx - 1], colors[gteIdx], q);
+        return Color.interpolateHsl(colors[gteIdx - 1], colors[gteIdx], q);
     };
 }